Refactor: pages/movie-list > features/movie-list로 변경
